+++ lib/grpc.dart
@@ -2,8 +2,19 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/client.dart';
-export 'src/server.dart';
-export 'src/shared.dart';
-export 'src/status.dart';
-export 'src/streams.dart';
+export 'src/client/call.dart';
+export 'src/client/channel.dart';
+export 'src/client/client.dart';
+export 'src/client/common.dart';
+export 'src/client/connection.dart';
+export 'src/client/options.dart';
+
+export 'src/server/call.dart';
+export 'src/server/handler.dart';
+export 'src/server/server.dart';
+export 'src/server/service.dart';
+
+export 'src/shared/security.dart';
+export 'src/shared/status.dart';
+export 'src/shared/streams.dart';
+export 'src/shared/timeout.dart';
