+++ lib/src/client/channel.dart
@@ -1,0 +1,71 @@
+// Copyright (c) 2017, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'dart:async';
+
+import '../shared/status.dart';
+
+import 'call.dart';
+import 'client.dart';
+import 'connection.dart';
+import 'options.dart';
+
+
+/// A channel to a virtual RPC endpoint.
+///
+/// For each RPC, the channel picks a [ClientConnection] to dispatch the call.
+/// RPCs on the same channel may be sent to different connections, depending on
+/// load balancing settings.
+class ClientChannel {
+  final String host;
+  final int port;
+  final ChannelOptions options;
+
+  // TODO(jakobr): Multiple connections, load balancing.
+  ClientConnection _connection;
+
+  bool _isShutdown = false;
+
+  ClientChannel(this.host,
+      {this.port = 443, this.options = const ChannelOptions.secure()});
+
+  /// Shuts down this channel.
+  ///
+  /// No further RPCs can be made on this channel. RPCs already in progress will
+  /// be allowed to complete.
+  Future<Null> shutdown() {
+    if (_isShutdown) return new Future.value();
+    _isShutdown = true;
+    return _connection.shutdown();
+  }
+
+  /// Terminates this channel.
+  ///
+  /// RPCs already in progress will be terminated. No further RPCs can be made
+  /// on this channel.
+  Future<Null> terminate() {
+    _isShutdown = true;
+    return _connection.terminate();
+  }
+
+  /// Returns a connection to this [Channel]'s RPC endpoint.
+  ///
+  /// The connection may be shared between multiple RPCs.
+  Future<ClientConnection> getConnection() async {
+    if (_isShutdown) throw new GrpcError.unavailable('Channel shutting down.');
+    return _connection ??= new ClientConnection(host, port, options);
+  }
+
+  /// Initiates a new RPC on this connection.
+  ClientCall<Q, R> createCall<Q, R>(
+      ClientMethod<Q, R> method, Stream<Q> requests, CallOptions options) {
+    final call = new ClientCall(method, requests, options);
+    getConnection().then((connection) {
+      if (call.isCancelled) return;
+      connection.dispatchCall(call);
+    }, onError: call.onConnectionError);
+    return call;
+  }
+}
+
