+++ lib/src/shared/streams.dart
