+++ test/server_test.dart
@@ -5,7 +5,6 @@
 import 'dart:async';
 
 import 'package:grpc/grpc.dart';
-import 'package:grpc/src/status.dart';
 import 'package:http2/transport.dart';
 import 'package:test/test.dart';
