+++ test/src/utils.dart
@@ -4,7 +4,7 @@
 
 import 'dart:convert';
 
-import 'package:grpc/src/streams.dart';
+import 'package:grpc/src/shared/streams.dart';
 import 'package:http2/transport.dart';
 import 'package:test/test.dart';
