+++ test/timeout_test.dart
@@ -9,8 +9,6 @@ import 'package:http2/transport.dart';
 import 'package:test/src/backend/invoker.dart';
 import 'package:test/test.dart';
 
-import 'package:grpc/src/client.dart';
-
 import 'src/client_utils.dart';
 import 'src/server_utils.dart';
 import 'src/utils.dart';
